+++ generators/view/index.js
@@ -63,7 +63,11 @@
 		}
 
 		inputIsThisYou() {
-			return isThisYou.set(this);
+			if (this.config.get('authorFirstName')) {
+				return isThisYou.set(this);
+			}
+			this.isThisYou = false;
+			return null;
 		}
 
 		inputAuthor() {