+++ generators/view/writing/jsFiles.js
@@ -13,6 +13,8 @@
 	const utils = require('../../common/utils.js');
 
 	module.exports = $that => {
+		const viewsPath = utils.getViewsPath($that);
+		$that.log(viewsPath);
 		const filePrefix = 'app/views/' + $that.viewNameCamel + '/' + $that.viewNameCamel + '.';
 		const config     = {
 			viewNameCamel  : $that.viewNameCamel,
@@ -26,19 +28,14 @@
 			nowTime        : $that.nowTime,
 			authorShort    : $that.authorShort || $that.config.get('authorShort'),
 			stateUrl       : $that.stateUrl,
-			stateName      : $that.stateName
+			stateName      : $that.stateName,
+			stateAbstract  : $that.stateAbstract
 		};
 
-		utils.copyFileTpl($that, 'controller.js', config, filePrefix + 'controller.js');
-		utils.copyFileTpl($that, 'service.js', config, filePrefix + 'service.js');
-		utils.copyFileTpl($that, 'constant.js', config, filePrefix + 'constant.js');
-
-		if ($that.stateAbstract) {
-			utils.copyFileTpl($that, 'states/state.abstract.js', config, filePrefix + 'state.js');
-		}
-		else {
-			utils.copyFileTpl($that, 'states/state.default.js', config, filePrefix + 'state.js');
-		}
+		utils.copyFileTpl($that, 'controller.txt', config, filePrefix + 'controller.js');
+		utils.copyFileTpl($that, 'service.txt', config, filePrefix + 'service.js');
+		utils.copyFileTpl($that, 'constant.txt', config, filePrefix + 'constant.js');
+		utils.copyFileTpl($that, 'state.txt', config, filePrefix + 'state.js');
 	};
 
 })();